RAG_chatbot/requirements.txt
torch>=2.0.0
# Dung version transformers on dinh, chua bat buoc torch>=2.6
transformers==4.44.2
FlagEmbedding>=1.2.10
pymongo>=4.10.0
python-dotenv>=1.0.1
numpy>=1.26.0
google-generativeai>=0.8.0
# Fix lỗi protobuf với transformers
protobuf>=4.21.0,<6.0.0
# Flask API
flask>=2.3.0
flask-cors>=4.0.0


